feat: api scan options under profile
