+++ app/adapters/api-scan-options.js
@@ -8,7 +8,7 @@ export default DRFAdapter.extend(DataAdapterMixin, {
   addTrailingSlashes: false,
   authorizer: 'authorizer:irene',
   urlForQueryRecord: function (q, modelName) {
-    let url = `${this.host}/${this.get('namespace')}/profiles/${q.id}/device_preference`;
+    let url = `${this.host}/${this.get('namespace')}/profiles/${q.id}/api_scan_options`;
     return url;
   }
 });
